+++ __tests__/components/Header-test.tsx
@@ -21,4 +21,25 @@ describe("Header", () => {
     expect(logout.find(".logoutIcon").length).to.equal(1);
     expect(logout.find("span").text()).to.contain("Log Out");
   });
+  it("should not render a nav element", () => {
+    expect(wrapper.find("nav").length).to.equal(0);
+  });
+  it("should render a nav element", () => {
+    wrapper = Enzyme.shallow(
+      <Header
+        text="Test Header!"
+        loggedIn={true}
+        nav={
+          <ul>
+            <li><a href="/about">About</a></li>
+            <li><a href="/account">account</a></li>
+            <li><a href="/misc">misc</a></li>
+          </ul>
+        }
+      />
+    );
+
+    expect(wrapper.find("nav").length).to.equal(1);
+    expect(wrapper.find("nav ul").length).to.equal(1);
+  });
 });