+++ src/components/Header.tsx
@@ -7,6 +7,8 @@ export interface HeaderProps {
   alt?: string;
   logOut?: string;
   loggedIn?: boolean;
+  /** This is expected to be a ul element or a div with an ul element. */
+  nav?: JSX.Element;
 }
 
 export default class Header extends React.Component<HeaderProps, {}> {
@@ -15,22 +17,24 @@ export default class Header extends React.Component<HeaderProps, {}> {
   };
 
   render(): JSX.Element {
-    let src = this.props.imgSrc ? require(`${this.props.imgSrc}`) : "";
+    const { imgSrc, alt, text, nav, loggedIn, logOut } = this.props;
+    let src = imgSrc ? require(`${imgSrc}`) : "";
     return(
       <header className="reusable-header clearfix">
-        <span>{this.props.text}</span>
         {
-          this.props.imgSrc &&
-          <img className="header-brand img-rounded" alt={this.props.alt || ""} src={src} />
+          imgSrc &&
+          <img className="header-brand img-rounded" alt={alt || ""} src={src} />
         }
-        { this.props.loggedIn &&
+        <span>{text}</span>
+        { loggedIn &&
           <a
             className="btn header-btn big inverted"
-            href={this.props.logOut}
+            href={logOut}
           >
             <span>Log Out <LogoutIcon /></span>
           </a>
         }
+        { nav && <nav role="navigation">{nav}</nav>}
       </header>
     );
   }